models.py
class Category:
    total_categories = 0
    total_unique_products = set()

    def __init__(self, name: str, description: str, products: list):
        self.name = name
        self.description = description
        self.products = products
        Category.total_categories += 1
        for product in products:
            Category.total_unique_products.add(product.name)

class Product:
    def __init__(self, name: str, description: str, price: float, quantity: int):
        self.name = name
        self.description = description
        self.price = price
        self.quantity = quantity
